requirement.txt
transformers==4.39.3
pytorch == 2.1.2
accelerate == 0.28.0